windows/certificates: Show Personal stores certificates
The normal Windows certificate store enumeration and opening APIs do not
return certificates in the `Personal` stores, which appear empty. These
certificates are actually stores on disk; not in the registry like most
others. Given this, we can manually inspect disk and retrieve these
certificates, which are located in a known directory location that has
been consistent since at least Windows 7.

The certificates are stored in a simple, but proprietary blob format.
This commit includes a small parser for it which extracts the
certificates, which can then be passed to the typical API
(CertCreateCertificateContext) for extracting information from a
certificate.